requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.6
# To update, run:
#
#    pip-compile requirements/requirements.in
#
asgiref==3.4.1
    # via django
black==22.1.0
    # via -r requirements/requirements.in
click==8.0.4
    # via
    #   black
    #   pip-tools
coverage==6.2
    # via -r requirements/requirements.in
dataclasses==0.8
    # via black
django==3.2.12
    # via -r requirements/requirements.in
django-widget-tweaks==1.4.11
    # via -r requirements/requirements.in
importlib-metadata==4.8.3
    # via
    #   click
    #   pep517
isort==5.10.1
    # via -r requirements/requirements.in
mypy-extensions==0.4.3
    # via black
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via pip-tools
pip-tools==6.4.0
    # via -r requirements/requirements.in
platformdirs==2.4.0
    # via black
python-dateutil==2.8.2
    # via -r requirements/requirements.in
python-dotenv==0.19.2
    # via -r requirements/requirements.in
pytz==2021.3
    # via django
six==1.16.0
    # via python-dateutil
sqlparse==0.4.2
    # via django
tomli==1.2.3
    # via
    #   black
    #   pep517
typed-ast==1.5.2
    # via black
typing-extensions==4.1.1
    # via
    #   asgiref
    #   black
    #   importlib-metadata
wheel==0.37.1
    # via pip-tools
zipp==3.6.0
    # via
    #   importlib-metadata
    #   pep517

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
